requirements.txt
torch==2.3.0
torchvision==0.18.0
transformers==4.41.1
accelerate==0.30.1
bitsandbytes==0.43.1
mysql-connector-python==8.3.0